Change class name of electro-thermal loading
This update changes the controller class name for the electro-thermal loading. The name is converted to "ControllerElectroThermalFolding" from "ControllerThermalFolding", which should represents better. This update on names open up naming spaces for a new type of loading where we change the ambient temperature to actuate the origami structure. Related codes are also updated. (Bascially, it is just for fixing some bad names I used previously, which unfortunately always happen to me in coding LOL)