Major update: Fixing ambiguity in mountain valley assignment
The previous code has ambiguity issue when assigning mountain valley fold (eg. setting a target fold angle = pi + a will randomly feed out mountain or valley depending on the way a pattern is input into the code environment). Now, the the new version has fixed it. A target folding angle of pi+a will consistently give an mountain fold and a target folding angle of pi-a will consistently give a valley fold. In addition to this major change, some small variable names are also updated. The examples are also updated accordingly.