Add dynamic electro-thermal loading process
This is a major update. A new solver is added to capture the transient dynamic electro-thermal loading behaviors of active origami. I will also work on a couple update recently to sort out the code and variable names. I have added a lot of functions into the package in the past year or so, but I have not clean the code. I will try to work on them in the coming weeks.