Changed Example 5 in mech examples
I changed this example because it would not work in octave. There
were some funny things happening when parsing the node locations, and
Octave doesn't support string comparisons of different lenghts using the
`==` operator. If you would like to add this to your library the `strcmp`
function also works in matlab.